Add thermal label PPTX generation

+++ features/table/immutable/src/desktopMain/kotlin/ru/pavlig43/immutable/internal/component/items/batch/BatchTableComponent.kt
@@ -43,7 +43,7 @@ private fun BatchWithBalanceOut.toUi(): BatchTableUi {
     return BatchTableUi(
         composeId = batchId,
         batchId = batchId,
-        balance = DecimalData3(balance.toLong()),
+        balance = DecimalData3(balance),
         productName = productName,
         vendorName = vendorName,
         dateBorn = dateBorn
